Init project + user interface structure
Files to init the project and user interface with tabs, containers and some widgets for adding a transaction.
The project has a MVC structure with 2 models: Transaction and Tag.
I can get user data for a transaction (I must verify user inputs), but there is no yet an ORM to save info in DB.