Refactoring + 1st code to display saved tags into tag field of transaction tab
Updated controller.py, containers.py and models.py in order to save tags and can display them into tag field of transaction tab.
For the moment, the tags can't be displayed.
Also, there is a refactor to display another container for 'update transaction' even if this one is the same as 'add transaction'.